requirements.txt
quixstreams
numpy
streamlit
kafka-python